Rename files and update main.js
