+++ nginx/www/dev/scripts/main.js
@@ -2,7 +2,7 @@ import { fetcher } from "./modules/fetcher.js";
 import * as bootstrap from "./bootstrap/bootstrap.bundle.min.js";
 import { initAuth } from "./modules/auth.js";
 import { createButton } from "./modules/buttonNav.js";
-import { watchProfile } from "./modules/profile.js";
+import { initSettings } from "./modules/settings.js";
 import { initRouter } from "./modules/router.js";
 import { createNavBar } from "./modules/navbar.js";
 import { generateModal } from "./modules/modal.js";
@@ -13,5 +13,5 @@ import { generateModal } from "./modules/modal.js";
 	await createButton();
 	initAuth();
 	initRouter();
-	watchProfile();
+	initSettings();
 })();
